fix tab bar bug and create UI components
